Moved TF / TF Agents dependencies out of specs.
learner.brains.specs is used by the API service as the learner
service. This moves TensorFlow and TensorFlow agents dependencies
from the specs module to tfa_specs. This makes it possible to use
the spec parsing and validation from specs in the API without
TensorFlow / TF-Agents, and data conversion functionality via
tfa_specs in the learner.
